+++ test/clients/dapp-client.spec.ts
@@ -37,6 +37,7 @@ import { MockTransport } from '../test-utils/MockTransport'
 import { availableTransports } from '../../src/utils/available-transports'
 import { ExposedPromise } from '../../src/utils/exposed-promise'
 import { Logger } from '../../src/utils/Logger'
+import { windowRef } from '../../src/MockWindow'
 
 // use chai-as-promised plugin
 chai.use(chaiAsPromised)
@@ -127,6 +128,7 @@ describe(`DAppClient`, () => {
 
   beforeEach(() => {
     sinon.restore()
+    ;(windowRef as any).beaconCreatedClientInstance = false
   })
 
   it(`should throw an error if initialized with an empty object`, async () => {
@@ -148,6 +150,10 @@ describe(`DAppClient`, () => {
     const storageStub = sinon.stub(storage, 'get').resolves()
 
     const message: PermissionResponse = {
+      appMetadata: {
+        senderId: 'sender-id',
+        name: 'test-wallet'
+      },
       id: 'some-id',
       version: BEACON_VERSION,
       senderId: 'sender-id',
@@ -319,6 +325,8 @@ describe(`DAppClient`, () => {
       connectedAt: new Date().getTime()
     }
 
+    const getPeersStub = sinon.stub(DAppClient.prototype, <any>'getPeer').resolves(peer1)
+
     await dAppClient.setActiveAccount(account1)
 
     activeAccount = await dAppClient.getActiveAccount()
@@ -328,6 +336,8 @@ describe(`DAppClient`, () => {
 
     activeAccount = await dAppClient.getActiveAccount()
     expect(activeAccount).to.be.undefined
+
+    expect(getPeersStub.callCount, 'getPeersStub').to.equal(1)
   })
 
   it(`should get app metadata`, async () => {
@@ -362,6 +372,8 @@ describe(`DAppClient`, () => {
   it(`should remove an account and unset active account`, async () => {
     const dAppClient = new DAppClient({ name: 'Test', storage: new LocalStorage() })
 
+    const getPeersStub = sinon.stub(DAppClient.prototype, <any>'getPeer').resolves(peer1)
+
     await (<any>dAppClient).accountManager.addAccount(account1)
     await (<any>dAppClient).accountManager.addAccount(account2)
     await dAppClient.setActiveAccount(account1)
@@ -373,10 +385,14 @@ describe(`DAppClient`, () => {
 
     expect(await dAppClient.getAccounts()).to.deep.equal([account2])
     expect(await dAppClient.getActiveAccount()).to.be.undefined
+
+    expect(getPeersStub.callCount, 'getPeersStub').to.equal(1)
   })
 
   it(`should remove an account and not unset active account`, async () => {
     const dAppClient = new DAppClient({ name: 'Test', storage: new LocalStorage() })
+
+    const getPeersStub = sinon.stub(DAppClient.prototype, <any>'getPeer').resolves(peer1)
 
     await (<any>dAppClient).accountManager.addAccount(account1)
     await (<any>dAppClient).accountManager.addAccount(account2)
@@ -389,10 +405,14 @@ describe(`DAppClient`, () => {
 
     expect(await dAppClient.getAccounts()).to.deep.equal([account1])
     expect(await dAppClient.getActiveAccount()).to.deep.equal(account1)
+
+    expect(getPeersStub.callCount, 'getPeersStub').to.equal(1)
   })
 
   it(`should remove all accounts and unset active account`, async () => {
     const dAppClient = new DAppClient({ name: 'Test', storage: new LocalStorage() })
+
+    const getPeersStub = sinon.stub(DAppClient.prototype, <any>'getPeer').resolves(peer1)
 
     await (<any>dAppClient).accountManager.addAccount(account1)
     await (<any>dAppClient).accountManager.addAccount(account2)
@@ -405,6 +425,8 @@ describe(`DAppClient`, () => {
 
     expect(await dAppClient.getAccounts()).to.deep.equal([])
     expect(await dAppClient.getActiveAccount()).to.deep.equal(undefined)
+
+    expect(getPeersStub.callCount, 'getPeersStub').to.equal(1)
   })
 
   it(`should remove peer and all its accounts`, async () => {
@@ -467,13 +489,13 @@ describe(`DAppClient`, () => {
       await dAppClient.checkPermissions(BeaconMessageType.OperationRequest)
       throw new Error('Should have failed')
     } catch (e) {
-      expect(eventsStub.callCount).to.equal(3)
       expect(eventsStub.firstCall.args[0]).to.equal(BeaconEvent.ACTIVE_TRANSPORT_SET) // Called in the constructor
       expect(eventsStub.firstCall.args[1]).to.equal(undefined)
       expect(eventsStub.secondCall.args[0]).to.equal(BeaconEvent.INTERNAL_ERROR)
       expect(eventsStub.secondCall.args[1]).to.equal('No active account set!')
-      expect(eventsStub.thirdCall.args[0]).to.equal(BeaconEvent.ACTIVE_ACCOUNT_SET) // Called in the constructor
-      expect(eventsStub.thirdCall.args[1]).to.equal(undefined)
+      // expect(eventsStub.thirdCall.args[0]).to.equal(BeaconEvent.ACTIVE_ACCOUNT_SET) // Called in the constructor
+      // expect(eventsStub.thirdCall.args[1]).to.equal(undefined)
+      expect(eventsStub.callCount).to.equal(2)
       expect(e.message).to.equal('No active account set!')
     }
   })
@@ -538,10 +560,14 @@ describe(`DAppClient`, () => {
     const dAppClient = new DAppClient({ name: 'Test', storage: new LocalStorage() })
 
     const permissionResponse: PermissionResponse = {
+      appMetadata: {
+        senderId: 'sender-id',
+        name: 'test-wallet'
+      },
       id: 'my-id',
       type: BeaconMessageType.PermissionResponse,
       version: BEACON_VERSION,
-      senderId: '69421294fd0136926639977666e8523550af4c126b6bcd429d3ae555c7aca3a3',
+      senderId: 'sender-id',
       publicKey: '444e1f4ab90c304a5ac003d367747aab63815f583ff2330ce159d12c1ecceba1',
       network: { type: NetworkType.MAINNET },
       scopes: [PermissionScope.SIGN, PermissionScope.OPERATION_REQUEST]
@@ -555,8 +581,24 @@ describe(`DAppClient`, () => {
       .stub(dAppClient, <any>'makeRequest')
       .resolves({ message: permissionResponse, connectionInfo })
 
+    const notifySuccessStub = sinon.stub(dAppClient, <any>'notifySuccess').resolves()
+
+    const getPeersStub = sinon.stub(Transport.prototype, 'getPeers').resolves([
+      {
+        id: '',
+        name: '',
+        publicKey: '69421294fd0136926639977666e8523550af4c126b6bcd429d3ae555c7aca3a3',
+        senderId: 'sender-id',
+        version: BEACON_VERSION,
+        type: 'p2p-pairing-request',
+        relayServer: ''
+      }
+    ])
+
     const response = await dAppClient.requestPermissions()
 
+    expect(getPeersStub.callCount, 'getPeersStub').to.equal(4)
+    expect(notifySuccessStub.callCount, 'notifySuccessStub').to.equal(1)
     expect(makeRequestStub.callCount).to.equal(1)
     expect(makeRequestStub.firstCall.args[0]).to.deep.equal({
       appMetadata: {
@@ -568,13 +610,20 @@ describe(`DAppClient`, () => {
       network: { type: 'mainnet' },
       scopes: ['operation_request', 'sign']
     })
+    delete response.accountInfo
     expect(response).to.deep.equal({
-      senderId: '69421294fd0136926639977666e8523550af4c126b6bcd429d3ae555c7aca3a3',
+      appMetadata: {
+        senderId: 'sender-id',
+        name: 'test-wallet'
+      },
+      id: 'my-id',
+      type: BeaconMessageType.PermissionResponse,
+      senderId: 'sender-id',
       address: 'tz1d75oB6T4zUMexzkr5WscGktZ1Nss1JrT7',
       network: { type: 'mainnet' },
       scopes: ['sign', 'operation_request'],
       publicKey: '444e1f4ab90c304a5ac003d367747aab63815f583ff2330ce159d12c1ecceba1',
-      threshold: undefined
+      version: '2'
     })
   })
 
@@ -583,7 +632,7 @@ describe(`DAppClient`, () => {
 
     const account: AccountInfo = {
       accountIdentifier: 'yQxM85PrJ718CA1N6oz',
-      senderId: '69421294fd0136926639977666e8523550af4c126b6bcd429d3ae555c7aca3a3',
+      senderId: 'sender-id',
       origin: {
         type: Origin.P2P,
         id: '69421294fd0136926639977666e8523550af4c126b6bcd429d3ae555c7aca3a3'
@@ -595,6 +644,8 @@ describe(`DAppClient`, () => {
       threshold: undefined,
       connectedAt: 1599142450653
     }
+
+    const getPeerStub = sinon.stub(DAppClient.prototype, <any>'getPeer').resolves(peer1)
 
     await (<any>dAppClient).accountManager.addAccount(account)
     await dAppClient.setActiveAccount(account)
@@ -616,8 +667,12 @@ describe(`DAppClient`, () => {
       .stub(dAppClient, <any>'makeRequest')
       .resolves({ message: signPayloadResponse, connectionInfo })
 
+    const notifySuccessStub = sinon.stub(dAppClient, <any>'notifySuccess').resolves()
+
     const response = await dAppClient.requestSignPayload({ payload: 'test-payload' })
 
+    expect(getPeerStub.callCount, 'getPeersStub').to.equal(2)
+    expect(notifySuccessStub.callCount, 'notifySuccessStub').to.equal(1)
     expect(makeRequestStub.callCount).to.equal(1)
     expect(makeRequestStub.firstCall.args[0]).to.deep.equal({
       type: BeaconMessageType.SignPayloadRequest,
@@ -626,9 +681,12 @@ describe(`DAppClient`, () => {
       sourceAddress: 'tz1d75oB6T4zUMexzkr5WscGktZ1Nss1JrT7'
     })
     expect(response).to.deep.equal({
+      id: 'my-id',
       senderId: 'sender-id',
       signingType: SigningType.RAW,
-      signature: 'my-signature'
+      signature: 'my-signature',
+      type: BeaconMessageType.SignPayloadResponse,
+      version: '2'
     })
   })
 
@@ -637,7 +695,7 @@ describe(`DAppClient`, () => {
 
     const account: AccountInfo = {
       accountIdentifier: 'yQxM85PrJ718CA1N6oz',
-      senderId: '69421294fd0136926639977666e8523550af4c126b6bcd429d3ae555c7aca3a3',
+      senderId: 'sender-id',
       origin: {
         type: Origin.P2P,
         id: '69421294fd0136926639977666e8523550af4c126b6bcd429d3ae555c7aca3a3'
@@ -649,6 +707,8 @@ describe(`DAppClient`, () => {
       threshold: undefined,
       connectedAt: 1599142450653
     }
+
+    const getPeerStub = sinon.stub(DAppClient.prototype, <any>'getPeer').resolves(peer1)
 
     await (<any>dAppClient).accountManager.addAccount(account)
     await dAppClient.setActiveAccount(account)
@@ -669,6 +729,8 @@ describe(`DAppClient`, () => {
       .stub(dAppClient, <any>'makeRequest')
       .resolves({ message: operationResponse, connectionInfo })
 
+    const notifySuccessStub = sinon.stub(dAppClient, <any>'notifySuccess').resolves()
+
     const operationDetails: PartialTezosOperation[] = [
       {
         kind: TezosOperationType.TRANSACTION,
@@ -678,6 +740,8 @@ describe(`DAppClient`, () => {
     ]
     const response = await dAppClient.requestOperation({ operationDetails })
 
+    expect(getPeerStub.callCount, 'getPeersStub').to.equal(2)
+    expect(notifySuccessStub.callCount, 'notifySuccessStub').to.equal(1)
     expect(makeRequestStub.callCount).to.equal(1)
     expect(makeRequestStub.firstCall.args[0]).to.deep.equal({
       type: BeaconMessageType.OperationRequest,
@@ -686,8 +750,11 @@ describe(`DAppClient`, () => {
       sourceAddress: 'tz1d75oB6T4zUMexzkr5WscGktZ1Nss1JrT7'
     })
     expect(response).to.deep.equal({
+      id: 'my-id',
       senderId: 'sender-id',
-      transactionHash: 'my-hash'
+      transactionHash: 'my-hash',
+      type: BeaconMessageType.OperationResponse,
+      version: '2'
     })
   })
 
@@ -696,7 +763,7 @@ describe(`DAppClient`, () => {
 
     const account: AccountInfo = {
       accountIdentifier: 'yQxM85PrJ718CA1N6oz',
-      senderId: '69421294fd0136926639977666e8523550af4c126b6bcd429d3ae555c7aca3a3',
+      senderId: 'sender-id',
       origin: {
         type: Origin.P2P,
         id: '69421294fd0136926639977666e8523550af4c126b6bcd429d3ae555c7aca3a3'
@@ -708,6 +775,8 @@ describe(`DAppClient`, () => {
       threshold: undefined,
       connectedAt: 1599142450653
     }
+
+    const getPeerStub = sinon.stub(DAppClient.prototype, <any>'getPeer').resolves(peer1)
 
     await (<any>dAppClient).accountManager.addAccount(account)
     await dAppClient.setActiveAccount(account)
@@ -728,8 +797,12 @@ describe(`DAppClient`, () => {
       .stub(dAppClient, <any>'makeRequest')
       .resolves({ message: broadcastResponse, connectionInfo })
 
+    const notifySuccessStub = sinon.stub(dAppClient, <any>'notifySuccess').resolves()
+
     const response = await dAppClient.requestBroadcast({ signedTransaction: 'signed-tx' })
 
+    expect(getPeerStub.callCount, 'getPeersStub').to.equal(2)
+    expect(notifySuccessStub.callCount, 'notifySuccessStub').to.equal(1)
     expect(makeRequestStub.callCount).to.equal(1)
     expect(makeRequestStub.firstCall.args[0]).to.deep.equal({
       type: BeaconMessageType.BroadcastRequest,
@@ -737,8 +810,11 @@ describe(`DAppClient`, () => {
       signedTransaction: 'signed-tx'
     })
     expect(response).to.deep.equal({
+      id: 'my-id',
       senderId: 'sender-id',
-      transactionHash: 'my-hash'
+      transactionHash: 'my-hash',
+      type: BeaconMessageType.BroadcastResponse,
+      version: '2'
     })
   })
 
@@ -751,7 +827,7 @@ describe(`DAppClient`, () => {
       await (<any>dAppClient).sendInternalError('some-message')
       throw new Error('Should not happen')
     } catch (e) {
-      expect(eventsStub.callCount).to.equal(2)
+      expect(eventsStub.callCount).to.equal(1)
       expect(eventsStub.firstCall.args[0]).to.equal(BeaconEvent.INTERNAL_ERROR)
       expect(eventsStub.firstCall.args[1]).to.equal('some-message')
       expect(e.message).to.equal('some-message')
@@ -804,15 +880,31 @@ describe(`DAppClient`, () => {
       errorType: BeaconErrorType.NOT_GRANTED_ERROR
     }
 
+    const walletInfoStub = sinon.stub(dAppClient, <any>'getWalletInfo').resolves({})
+
+    ;(<any>dAppClient)._activePeer = ExposedPromise.resolve({
+      id: '',
+      name: '',
+      publicKey: '69421294fd0136926639977666e8523550af4c126b6bcd429d3ae555c7aca3a3',
+      senderId: 'sender-id',
+      version: BEACON_VERSION,
+      type: 'p2p-pairing-request',
+      relayServer: ''
+    })
+
     try {
       await (<any>dAppClient).handleRequestError(request, error)
       throw new Error('Should not happen')
     } catch (e) {
-      expect(eventsStub.callCount).to.equal(2)
-      expect(eventsStub.firstCall.args[0]).to.equal(BeaconEvent.PERMISSION_REQUEST_ERROR)
-      expect(eventsStub.firstCall.args[1]).to.equal(error)
-      expect(eventsStub.secondCall.args[0]).to.equal(BeaconEvent.ACTIVE_TRANSPORT_SET)
-      expect(eventsStub.secondCall.args[1]).to.equal(undefined)
+      expect(walletInfoStub.callCount, 'walletInfoStub').to.equal(1)
+
+      expect(eventsStub.callCount).to.equal(3)
+      expect(eventsStub.getCall(0).args[0]).to.equal(BeaconEvent.ACTIVE_TRANSPORT_SET)
+      expect(eventsStub.getCall(0).args[1]).to.equal(undefined)
+      expect(eventsStub.getCall(1).args[0]).to.equal(BeaconEvent.PERMISSION_REQUEST_ERROR)
+      expect(eventsStub.getCall(1).args[1]).to.deep.eq({ errorResponse: error, walletInfo: {} })
+      expect(eventsStub.getCall(2).args[0]).to.equal(BeaconEvent.ACTIVE_ACCOUNT_SET)
+      expect(eventsStub.getCall(2).args[1]).to.equal(undefined)
       expect(e.description).to.equal(
         'You do not have the necessary permissions to perform this action. Please initiate another permission request and give the necessary permissions.'
       )
@@ -833,9 +925,7 @@ describe(`DAppClient`, () => {
       await (<any>dAppClient).handleRequestError(request, error)
       throw new Error('Should not happen')
     } catch (e) {
-      expect(eventsStub.callCount).to.equal(1)
-      expect(eventsStub.firstCall.args[0]).to.equal(BeaconEvent.ACTIVE_TRANSPORT_SET)
-      expect(eventsStub.firstCall.args[1]).to.equal(undefined)
+      expect(eventsStub.callCount).to.equal(0)
       expect(e).to.equal(error)
     }
   })
@@ -849,17 +939,17 @@ describe(`DAppClient`, () => {
     const response = {}
 
     await (<any>dAppClient).notifySuccess(request, response)
-    expect(eventsStub.callCount).to.equal(2)
+    expect(eventsStub.callCount).to.equal(1)
     expect(eventsStub.firstCall.args[0]).to.equal(BeaconEvent.PERMISSION_REQUEST_SUCCESS)
     expect(eventsStub.firstCall.args[1]).to.equal(response)
-    expect(eventsStub.secondCall.args[0]).to.equal(BeaconEvent.ACTIVE_TRANSPORT_SET)
-    expect(eventsStub.secondCall.args[1]).to.equal(undefined)
   })
 
   it(`should create a request`, async () => {
     return new Promise(async (resolve, _reject) => {
       const dAppClient = new DAppClient({ name: 'Test', storage: new LocalStorage() })
       const sendStub = sinon.stub(Transport.prototype, 'send').resolves()
+
+      const getPeerStub = sinon.stub(DAppClient.prototype, <any>'getPeer').resolves(peer1)
 
       await initClientWithMock(dAppClient)
 
@@ -878,6 +968,8 @@ describe(`DAppClient`, () => {
       const promise: Promise<any> = (<any>dAppClient).makeRequest(input)
 
       setTimeout(async () => {
+        expect(getPeerStub.callCount, 'getPeersStub').to.equal(1)
+
         expect(initStub.callCount, 'initStub').to.equal(1)
         expect(rateLimitStub.callCount, 'rateLimitStub').to.equal(1)
         expect(permissionStub.callCount, 'permissionStub').to.equal(1)
@@ -886,7 +978,7 @@ describe(`DAppClient`, () => {
         expect(addRequestStub.firstCall.args[1].isPending(), 'addRequestStub').to.be.true
         expect(sendStub.callCount, 'sendStub').to.equal(1)
         expect(sendStub.firstCall.args[0].length, 'sendStub').to.be.greaterThan(20)
-        expect(sendStub.firstCall.args[1], 'sendStub').to.be.undefined
+        expect(sendStub.firstCall.args[1], 'sendStub').to.equal(peer1)
         expect(eventsStub.callCount, 'eventsStub').to.equal(1)
         expect(typeof promise).to.equal('object')