+++ test/clients/wallet-client.spec.ts
@@ -24,6 +24,7 @@ import * as sinon from 'sinon'
 
 import { WalletClient } from '../../src/clients/wallet-client/WalletClient'
 import { ExtendedP2PPairingRequest } from '../../src/types/P2PPairingRequest'
+import { windowRef } from '../../src/MockWindow'
 
 // use chai-as-promised plugin
 chai.use(chaiAsPromised)
@@ -76,6 +77,7 @@ describe(`WalletClient`, () => {
 
   beforeEach(() => {
     sinon.restore()
+    ;(windowRef as any).beaconCreatedClientInstance = false
   })
 
   it(`should throw an error if initialized with an empty object`, async () => {
@@ -166,7 +168,8 @@ describe(`WalletClient`, () => {
       publicKey: '69421294fd0136926639977666e8523550af4c126b6bcd429d3ae555c7aca3a3'
     }
 
-    ;(<any>walletClient).pendingRequests = [{ id }]
+    ;(<any>walletClient).pendingRequests.push([{ id }, {}])
+
     const respondStub = sinon.stub(walletClient, <any>'respondToMessage').resolves()
     const appMetadataManagerStub = sinon
       .stub(AppMetadataManager.prototype, 'getAppMetadata')
@@ -304,7 +307,7 @@ describe(`WalletClient`, () => {
     // expect(sendStub.firstCall.args[0]).to.equal('aRNACa2rFgw2dfAugetVZpzSbMdahH')
   })
 
-  it.only(`should respond with an error message`, async () => {
+  it(`should respond with an error message`, async () => {
     const walletClient = new WalletClient({ name: 'Test', storage: new LocalStorage() })
 
     const respondStub = sinon.stub(walletClient, <any>'respondToMessage').resolves()
@@ -335,7 +338,7 @@ describe(`WalletClient`, () => {
     expect(respondStub.firstCall.args[0].errorData).not.to.be.undefined
   })
 
-  it.only(`should respond with an error message`, async () => {
+  it(`should respond with an error message`, async () => {
     const walletClient = new WalletClient({ name: 'Test', storage: new LocalStorage() })
 
     const respondStub = sinon.stub(walletClient, <any>'respondToMessage').resolves()
@@ -368,7 +371,7 @@ describe(`WalletClient`, () => {
     expect(respondStub.firstCall.args[0].errorData).to.be.undefined
   })
 
-  it.only(`should not respond with an invalid error message`, async () => {
+  it(`should not respond with an invalid error message`, async () => {
     const walletClient = new WalletClient({ name: 'Test', storage: new LocalStorage() })
 
     const respondStub = sinon.stub(walletClient, <any>'respondToMessage').resolves()